requirements.txt
future==0.16.0 ; python_version <= '2.7'
requests==2.12.4
robotframework==3.0.3
